Time: 0 ms (100.00%) | Memory: 7.9 MB (39.48%) - LeetSync

+++ 2288-count-operations-to-obtain-zero/count-operations-to-obtain-zero.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
-    long countOperations(int num1, int num2) {
-        long count=0;
+    int countOperations(int num1, int num2) {
+        int count=0;
         while(num1 && num2){
             if(num1<=num2){
                 num2=num2-num1;
